refactor: 프로젝트 구조 개선 및 컴포넌트 재구성
프로젝트의 전반적인 구조를 개선하고 컴포넌트를 재구성했습니다.

주요 변경사항:

1. Section 컴포넌트 통합 및 재배치
   - src/app/components/project/sidebar/* → src/app/components/section/
   - ProjectActions → ProjectActionsSection (일관성 개선)
   - 영역 구분 컴포넌트 파일명에 'Section' suffix 추가

2. 공용 컴포넌트를 /shared로 분리
   - ProjectGrid → src/shared/project/ProjectGrid.tsx
   - 여러 도메인에서 재사용되는 컴포넌트 중앙화

3. 사이드바 컴포넌트 재배치
   - ProjectSidebar → src/app/components/sidebar/
   - TeamSidebar → src/app/components/sidebar/
   - 사이드바 컴포넌트 통합 관리

4. 파일명 개선
   - MetaInfo → ProjectCardMetadata (의미 명확화)

5. 타입 정의 업데이트
   - database.types.ts에 is_official 필드 추가

6. Import 경로 일관성 개선
   - 절대 경로(@/) 사용으로 통일
   - 컴포넌트 재배치에 따른 모든 import 경로 업데이트

모든 변경사항은 lint/build 검증 완료
