+++ src/app/(box-layout)/team/Team.tsx
@@ -2,9 +2,9 @@
 import { getTeamData } from '@/services/team/getTeamData.server';
 import { getTeamProjects } from '@/services/team/getTeamProjects.server';
 import { notFound } from 'next/navigation';
-import TeamSidebar from './TeamSidebar';
+import TeamSidebar from '@/app/components/sidebar/TeamSidebar';
 import ProfileIcon from '@/app/components/card/portfolio/ProfileIcon';
-import ProjectGrid from '@/app/components/project/components/ProjectGrid';
+import ProjectGrid from '@/shared/project/ProjectGrid';
 
 const Team = async ({ teamName }: { teamName: string }) => {
   const teamDetail = (await getTeamData(teamName)) ?? notFound();
