+++ src/app/components/project/ProjectPage.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 import ProjectMainContent from '@/app/components/project/components/ProjectMainContent';
-import ProjectSidebar from '@/app/components/project/components/ProjectSidebar';
+import ProjectSidebar from '@/app/components/sidebar/ProjectSidebar';
 import { getProjectDetailViewModel } from '@/services/project/getProjectDetail.server';
 import checkProjectEditPermission from '@/services/project/checkProjectEditPermission.server';
