+++ src/app/components/section/ProjectLinkSection.tsx
@@ -32,4 +32,3 @@ export const ProjectLinkSection = ({
     </section>
   );
 };
-
