+++ src/app/components/section/ProjectSummarySection.tsx
