+++ src/app/components/section/ProjectTechnologiesSection.tsx
@@ -17,4 +17,3 @@ export const ProjectTechnologiesSection = ({
     )}
   </section>
 );
-
